+++ coreax/solvers/__init__.py
@@ -23,7 +23,7 @@ from coreax.solvers.base import (
 )
 from coreax.solvers.composite import CompositeSolver, MapReduce
 from coreax.solvers.coresubset import (
-    GreedyCMMD,
+    ConditionalKernelHerding,
     HerdingState,
     KernelHerding,
     RandomSample,
@@ -46,5 +46,5 @@ __all__ = [
     "SteinThinning",
     "MapReduce",
     "PaddingInvariantSolver",
-    "GreedyCMMD",
+    "ConditionalKernelHerding",
 ]